feat(web): Display change: Display notes according to missions
