+++ web/src/component/Button.tsx
@@ -1,15 +1,15 @@
 import { ButtonProps, Button, styled } from '@mui/material';
 
-var primaryColor = '#7c44f3';
+const primaryColor = '#7c44f3';
 
-export const PrimaryButton = styled(Button)<ButtonProps>(({ theme }) => ({
+export const PrimaryButton = styled(Button)<ButtonProps>(() => ({
     backgroundColor: primaryColor,
     '&:hover': {
         backgroundColor: '#632add',
     },
 }));
 
-export const SecondaryButton = styled(Button)<ButtonProps>(({ theme }) => ({
+export const SecondaryButton = styled(Button)<ButtonProps>(() => ({
     color: primaryColor,
     border: '1px solid',
     borderColor: primaryColor,
