+++ web/src/pages/Dashboard/DashBoardNote.type.tsx
@@ -1,4 +1,3 @@
-
 export interface IDashboardNotes {
     id: number;
     title: string;
